guests/linux: Centralize public_key logic
