+++ plugins/guests/linux/plugin.rb
@@ -22,8 +22,8 @@ module VagrantPlugins
       end
 
       guest_capability(:linux, :insert_public_key) do
-        require_relative "cap/insert_public_key"
-        Cap::InsertPublicKey
+        require_relative "cap/public_key"
+        Cap::PublicKey
       end
 
       guest_capability(:linux, :shell_expand_guest_path) do
@@ -68,8 +68,8 @@ module VagrantPlugins
       end
 
       guest_capability(:linux, :remove_public_key) do
-        require_relative "cap/remove_public_key"
-        Cap::RemovePublicKey
+        require_relative "cap/public_key"
+        Cap::PublicKey
       end
 
       guest_capability(:linux, :rsync_installed) do
